Refactor geo override database
Note that the Госуслуги certificate (https://www.gosuslugi.ru/tls)
will NOT need any special handling in this database as it is using
a ccTLD -- this means it'll be detected correctly by default. ☺︎

This database is ONLY for overriding the subset of certificate
authorities that chose to host their official business documents
on "vanity" domains.